+++ yash-syntax/src/parser/lex/core.rs
@@ -21,7 +21,7 @@ use super::op::Operator;
 use crate::alias::Alias;
 use crate::alias::EmptyGlossary;
 use crate::input::Context;
-use crate::input::Input;
+use crate::input::InputObject;
 use crate::input::Memory;
 use crate::parser::core::Result;
 use crate::parser::error::Error;
@@ -154,7 +154,7 @@ fn ex<I: IntoIterator<Item = SourceChar>>(i: I) -> impl Iterator<Item = SourceCh
 
 /// Core part of the lexical analyzer.
 struct LexerCore<'a> {
-    input: Box<dyn Input + 'a>,
+    input: Box<dyn InputObject + 'a>,
     state: InputState,
     raw_code: Rc<Code>,
     source: Vec<SourceCharEx>,
@@ -165,7 +165,7 @@ impl<'a> LexerCore<'a> {
     /// Creates a new lexer core that reads using the given input function.
     #[must_use]
     fn new(
-        input: Box<dyn Input + 'a>,
+        input: Box<dyn InputObject + 'a>,
         start_line_number: NonZeroU64,
         source: Rc<Source>,
     ) -> LexerCore<'a> {
@@ -454,7 +454,7 @@ impl<'a> Lexer<'a> {
     /// Creates a new lexer that reads using the given input function.
     #[must_use]
     pub fn new(
-        input: Box<dyn Input + 'a>,
+        input: Box<dyn InputObject + 'a>,
         start_line_number: NonZeroU64,
         source: Rc<Source>,
     ) -> Lexer<'a> {
@@ -838,6 +838,7 @@ impl<'a, 'b> DerefMut for WordLexer<'a, 'b> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::input::Input;
     use crate::parser::error::ErrorCause;
     use crate::parser::error::SyntaxError;
     use assert_matches::assert_matches;
@@ -867,7 +868,6 @@ mod tests {
             }
         }
         impl std::error::Error for Failing {}
-        #[async_trait::async_trait(?Send)]
         impl Input for Failing {
             async fn next_line(&mut self, _: &Context) -> crate::input::Result {
                 Err(std::io::Error::new(std::io::ErrorKind::Other, Failing))
@@ -892,7 +892,6 @@ mod tests {
         struct InputMock {
             first: bool,
         }
-        #[async_trait::async_trait(?Send)]
         impl Input for InputMock {
             async fn next_line(&mut self, context: &Context) -> crate::input::Result {
                 assert_eq!(context.is_first_line(), self.first);